Move mqtt publish into common
