+++ src/events/events.gateway.ts
@@ -6,12 +6,10 @@ import {
   WsResponse,
 } from "@nestjs/websockets";
 import { Injectable, UseGuards } from "@nestjs/common";
-import { uuid } from "systeminformation";
-import mqtt from "mqtt";
 import WebSocket from "ws";
 
 import { Event } from "./entities/event.entity";
-import { getMqttOptions } from "../common";
+import { mqttPublish } from "../common";
 import { startServer, stopServer } from "../main";
 import { WsAuthGuard } from "../wsAuth.guard";
 import logger from "../logger";
@@ -54,28 +52,7 @@ export class EventsGateway {
             )
           );
       }, 200);
-    const { enabled, host, port, username, password } = await getMqttOptions();
-    if (enabled) {
-      const id = (await uuid()).os;
-      const mqttClient = mqtt.connect(`mqtt://${host}:${port}`, {
-        username,
-        password,
-        clientId: id,
-      });
-      const topic = `systembridge/event/${id}/${data.name}`;
-      logger.debug(`WebSocket - MQTT - Publishing to topic ${topic}`);
-      mqttClient.publish(
-        topic,
-        JSON.stringify(data.data),
-        { qos: 1, retain: true },
-        (error?: Error) => {
-          if (error)
-            logger.error(
-              `WebSocket - MQTT - Error publishing message: ${error.message}`
-            );
-        }
-      );
-    }
+    await mqttPublish(`event/${data.name}`, JSON.stringify(data.data));
     return { event: "event-sent", data };
   }
 }
